// src/api/anonymous-review/models/anonymousReview.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * AnonymousReviewService
 * OpenAPI spec version: 1.0.0
 */
import type { ReviewStats } from "./reviewStats";

export interface AnonymousReview {
	id: number;
	comments: string;
	stats: ReviewStats;
	tags: string[];
	created_at: string;
}