// src/api/review/models/review.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * ReviewService
 * OpenAPI spec version: 1.0.0
 */
import type { ReviewStats } from "./reviewStats";

export interface Review {
	id: number;
	account_id: number;
	course_code: string;
	comments?: string;
	stats: ReviewStats;
	tags: string[];
	created_at: string;
}